Add request_sent? method

+++ app/helpers/application_helper.rb
@@ -23,13 +23,11 @@ module ApplicationHelper
 
   def request_sent?(user)
     request = Friendship.find_by(user_id: user.id, friend_id: current_user.id, confirmed: false)
-    if request
-      true
-    end
+    request
   end
 
   def send_or_recall_friend_request(user)
-    if current_user != user && !friends?(user)
+    if current_user != user && !friends?(user) && !request_sent?(user)
       request = Friendship.find_by(user_id: current_user.id, friend_id: user.id, confirmed: false)
       if request
         link_to('Recall friend request', user_friendship_path(id: request.id, user_id: current_user.id, friend_id: user.id), method: :delete, class: 'btn-alert')
